PongGame

+++ Day20-21-SnakeGame/food.py
@@ -4,6 +4,7 @@ import random
 WIDTH = 600
 HEIGHT = 600
 
+
 class Food(Turtle):
 
     def __init__(self):
@@ -12,9 +13,8 @@ class Food(Turtle):
         self.color("white")
         self.penup()
         self.refresh()
-    
 
     def refresh(self):
-        random_x = random.randint((WIDTH/2 - 20) * (-1), WIDTH/2 - 20)
-        random_y = random.randint((HEIGHT/2 - 20) * (-1), HEIGHT/2 - 20)
+        random_x = random.randint(int(WIDTH/2 - 20) * (-1), int(WIDTH/2) - 20)
+        random_y = random.randint(int(HEIGHT/2 - 20) * (-1), int(HEIGHT/2) - 20)
         self.goto(random_x, random_y)
